# frontend/kucukdevapi/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "kucukdevapi",
    "version": "1.0.0"
  },
  "paths": {
...
    "/universities/{unid}/semesters/{unisid}/lessons": {
      "get": {
        "tags": [
          "university lessons"
        ],
        "summary": "List University Lessons",
        "description": "list all lessons for a semesters of a university with given universityID and universitySemesterID",
        "operationId": "listUniversitySemesterLessons",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Unid",
              "type": "string"
            },
            "name": "unid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Unisid",
              "type": "string"
            },
            "name": "unisid",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "List all lessons of a university semester",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response List University Lessons Universities  Unid  Semesters  Unisid  Lessons Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UniversityLessonAPIModel"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "university lessons"
        ],
        "summary": "Create University Lesson",
        "description": "Create a lesson for a semester of a university with given universityID and universitySemesterID",
        "operationId": "createUniversityLesson",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Unid",
              "type": "string"
            },
            "name": "unid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Unisid",
              "type": "string"
            },
            "name": "unisid",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UniversityLessonModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Add new university lesson",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageCreate"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageCreate"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/universities/{unid}/semesters/current-semester/lessons/find-code": {
      "get": {
        "tags": [
          "university lessons"
        ],
        "summary": "Show Lesson With Code",
        "description": "Get a single lesson of a university semester with given universityID and Lesson Code",
        "operationId": "getSingleLessonWithCode",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Unid",
              "type": "string"
            },
            "name": "unid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            },
            "name": "code",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Get a single lessons with code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UniversityLessonAPIModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/universities/{unid}/semesters/{unisid}/lessons/{unilid}": {
      "get": {
        "tags": [
          "university lessons"
        ],
        "summary": "Show University Lesson",
        "description": "Get a single lesson of a university semester with given universityID, universitySemesterID and universityLessonID",
        "operationId": "getSingleUniversitySemesterLesson",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Unid",
              "type": "string"
            },
            "name": "unid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Unisid",
              "type": "string"
            },
            "name": "unisid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Unilid",
              "type": "string"
            },
            "name": "unilid",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Get a single lessons of a university semester",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UniversityLessonAPIModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "university lessons"
        ],
        "summary": "Update University Lesson",
        "description": "Update lesson of a university semester with given universityID, universitySemesterID and universityLessonID",
        "operationId": "updateUniversityLesson",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Unid",
              "type": "string"
            },
            "name": "unid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Unisid",
              "type": "string"
            },
            "name": "unisid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Unilid",
              "type": "string"
            },
            "name": "unilid",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UniversityLessonAPIModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Update a university lesson",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "university lessons"
        ],
        "summary": "Delete University Lesson",
        "description": "Delete a university lesson with given universityID, universitySemesterID and universityLessonID",
        "operationId": "deleteUniversityLesson",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Unid",
              "type": "string"
            },
            "name": "unid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Unisid",
              "type": "string"
            },
            "name": "unisid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Unilid",
              "type": "string"
            },
            "name": "unilid",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Delete university lesson",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "Message": {
        "title": "Message",
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          }
        }
      },
      "MessageCreate": {
        "title": "MessageCreate",
        "required": [
          "_id",
          "message"
        ],
        "type": "object",
        "properties": {
          "_id": {
            "title": " Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "example": {
          "_id": "61ddea901311ecaed99afb7c",
          "message": "string"
        }
      },
...
      "UniversityLessonAPIModel": {
        "title": "UniversityLessonAPIModel",
        "required": [
          "name",
          "code",
          "ects",
          "absenceLimit"
        ],
        "type": "object",
        "properties": {
          "_id": {
            "title": " Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "ects": {
            "title": "Ects",
            "minimum": 0,
            "type": "number"
          },
          "absenceLimit": {
            "title": "Absencelimit",
            "minimum": 0,
            "type": "integer"
          },
          "sections": {
            "title": "Sections",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UniversitySectionAPIModel"
            },
            "default": []
          }
        },
        "example": {
          "name": "ART OF COMPUTING",
          "code": "COMP101",
          "ects": 6,
          "absenceLimit": 0
        }
      },
      "UniversityLessonModel": {
        "title": "UniversityLessonModel",
        "required": [
          "name",
          "code",
          "ects",
          "absenceLimit",
          "section",
          "instructor",
          "slots"
        ],
        "type": "object",
        "properties": {
          "_id": {
            "title": " Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "ects": {
            "title": "Ects",
            "minimum": 0,
            "type": "number"
          },
          "absenceLimit": {
            "title": "Absencelimit",
            "minimum": 0,
            "type": "integer"
          },
          "section": {
            "title": "Section",
            "type": "string"
          },
          "instructor": {
            "title": "Instructor",
            "type": "string"
          },
          "slots": {
            "title": "Slots",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UniversitySlotModel"
            }
          }
        },
        "example": {
          "name": "ART OF COMPUTING",
          "code": "COMP101",
          "ects": 6,
          "absenceLimit": 8,
          "section": "01",
          "instructor": "JACK JOE",
          "slots": [
            {
              "day": 2,
              "hour": 7,
              "isLab": 0
            },
            {
              "day": 2,
              "hour": 8,
              "isLab": 0
            }
          ]
        }
      },
...
      "UniversitySectionAPIModel": {
        "title": "UniversitySectionAPIModel",
        "type": "object",
        "properties": {
          "_id": {
            "title": " Id",
            "type": "string"
          },
          "section": {
            "title": "Section",
            "type": "string"
          },
          "instructor": {
            "title": "Instructor",
            "type": "string"
          },
          "slots": {
            "title": "Slots",
            "type": "array",
            "items": {
              "type": "array",
              "items": {
                "type": "integer"
              }
            }
          }
        },
        "example": {
          "section": "01",
          "instructor": "JACK JOE",
          "slots": [
            [
              2,
              7,
              0
            ],
            [
              2,
              8,
              0
            ]
          ]
        }
      },
...
      "UniversitySlotModel": {
        "title": "UniversitySlotModel",
        "required": [
          "day",
          "hour",
          "isLab"
        ],
        "type": "object",
        "properties": {
          "day": {
            "title": "Day",
            "maximum": 4,
            "minimum": 0,
            "type": "integer"
          },
          "hour": {
            "title": "Hour",
            "maximum": 15,
            "minimum": 0,
            "type": "integer"
          },
          "isLab": {
            "title": "Islab",
            "maximum": 1,
            "minimum": 0,
            "type": "integer"
          }
        }
      },
...
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ]
}